+++ src/main/java/se/uog/application/AppModel.java
@@ -1,16 +1,24 @@
 package se.uog.application;
 
-import se.uog.course.*;
+import java.beans.PropertyChangeSupport;
+
+import javax.swing.DefaultListModel;
+
+import se.uog.course.AdminCourseTableModel;
+import se.uog.course.CDCourseTableModel;
+import se.uog.course.Course;
+import se.uog.course.CourseTableModel;
+import se.uog.course.PTTCourseTableModel;
+
 import se.uog.qualification.Qualification;
 import se.uog.qualification.QualificationTableModel;
+
 import se.uog.teacher.Teacher;
 import se.uog.teacher.TeacherTableModel;
 import se.uog.training.Training;
 import se.uog.training.TrainingTableModel;
-import se.uog.user.User;
 
-import javax.swing.*;
-import java.beans.PropertyChangeSupport;
+import se.uog.user.User;
 
 /**
  * The AppModel class, a wrapper containing the Core Object lists consumed by
