+++ src/main/java/se/uog/application/HomePage.java
@@ -1,9 +1,14 @@
 package se.uog.application;
 
-import se.uog.user.User;
+import java.awt.GridBagConstraints;
+import java.awt.GridBagLayout;
 
-import javax.swing.*;
-import java.awt.*;
+import javax.swing.BorderFactory;
+import javax.swing.JButton;
+import javax.swing.JLabel;
+import javax.swing.JPanel;
+
+import se.uog.user.User;
 
 @SuppressWarnings("serial")
 public class HomePage extends JPanel {
