+++ src/main/java/se/uog/application/TablePageView.java
@@ -1,9 +1,8 @@
-package se.uog;
+package se.uog.application;
 
+import javax.swing.JPanel;
 import se.uog.table.JObjectTable;
 import se.uog.table.ObjectTableModel;
-
-import javax.swing.*;
 
 @SuppressWarnings("serial")
 public class TablePageView<E> extends JPanel {
