+++ src/main/java/se/uog/course/AdminCourseTableModel.java
@@ -1,16 +1,19 @@
 package se.uog.course;
 
-import se.uog.qualification.Qualification;
-import se.uog.table.ObjectTableColumn;
-import se.uog.table.ObjectTableColumnBuilder;
-import se.uog.table.ObjectTableListSelector;
-import se.uog.teacher.Teacher;
-
-import javax.swing.*;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 import java.util.function.BiFunction;
+
+import javax.swing.DefaultListModel;
+
+import se.uog.qualification.Qualification;
+
+import se.uog.table.ObjectTableColumn;
+import se.uog.table.ObjectTableColumnBuilder;
+import se.uog.table.ObjectTableListSelector;
+
+import se.uog.teacher.Teacher;
 
 public class AdminCourseTableModel extends CourseTableModel {
