+++ src/main/java/se/uog/course/Course.java
@@ -1,10 +1,10 @@
 package se.uog.course;
 
-import se.uog.qualification.Qualification;
-import se.uog.teacher.Teacher;
-
 import java.util.ArrayList;
 import java.util.List;
+
+import se.uog.qualification.Qualification;
+import se.uog.teacher.Teacher;
 
 public class Course {
     private String courseDirectorName;
