+++ src/main/java/se/uog/course/PTTCourseTableModel.java
@@ -1,11 +1,12 @@
 package se.uog.course;
 
-import se.uog.table.ObjectTableColumn;
-import se.uog.table.ObjectTableColumnBuilder;
-
-import javax.swing.*;
 import java.util.ArrayList;
 import java.util.List;
+
+import javax.swing.DefaultListModel;
+
+import se.uog.table.ObjectTableColumn;
+import se.uog.table.ObjectTableColumnBuilder;
 
 public class PTTCourseTableModel extends CourseTableModel {
