+++ src/main/java/se/uog/database/AppModelDeserializer.java
@@ -1,14 +1,21 @@
 package se.uog.database;
 
-import com.google.gson.*;
+import java.lang.reflect.Type;
+
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.gson.JsonDeserializationContext;
+import com.google.gson.JsonDeserializer;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParseException;
+
 import se.uog.IDReferenced;
 import se.uog.application.AppModel;
 import se.uog.course.Course;
 import se.uog.qualification.Qualification;
 import se.uog.teacher.Teacher;
 import se.uog.training.Training;
-
-import java.lang.reflect.Type;
 
 /**
  * A custom deserializer for the AppModel wrapper.
