+++ src/main/java/se/uog/logger/Logger.java
@@ -1,6 +1,7 @@
 package se.uog.logger;
 
 import java.io.IOException;
+
 import java.util.logging.FileHandler;
 import java.util.logging.Formatter;
 import java.util.logging.Handler;
