+++ src/main/java/se/uog/table/JObjectTable.java
@@ -1,12 +1,19 @@
 package se.uog.table;
 
-import javax.swing.*;
-import javax.swing.table.TableCellEditor;
-import javax.swing.table.TableColumnModel;
-import java.awt.*;
+import java.awt.BorderLayout;
+import java.awt.Dimension;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
+
 import java.util.List;
+
+import javax.swing.DefaultListModel;
+import javax.swing.JButton;
+import javax.swing.JPanel;
+import javax.swing.JScrollPane;
+import javax.swing.JTable;
+import javax.swing.table.TableCellEditor;
+import javax.swing.table.TableColumnModel;
 
 /**
  * A class that encapsulates an ObjectTable component. It was designed to allow the class to be
