+++ src/main/java/se/uog/table/ObjectTableColumn.java
@@ -1,14 +1,13 @@
 package se.uog.table;
 
-import javax.swing.table.TableCellEditor;
 import java.util.function.BiConsumer;
 import java.util.function.Function;
+import javax.swing.table.TableCellEditor;
 
 /**
  * This class is consumed by {@link se.uog.table.ObjectTableModelAdaptor} and
- * {@link se.uog.table.ObjectTable}. It tells the model and view how to map each
- * table column to the underlying class's attribute, including it's getters and
- * setters.
+ * {@link se.uog.table.ObjectTable}. It tells the model and view how to map each table column to the
+ * underlying class's attribute, including it's getters and setters.
  * <p>
  * Getters are protected as they are only consumed by this package's classes.
  *
@@ -27,26 +26,24 @@ public class ObjectTableColumn<E> {
      * An object which maps an element's attribute to an ObjectTable column.
      *
      * @param columnTitle      The title to display above the column.
-     * @param columnClass      The class type the column deals with. This allows
-     *                         automatic setting of renderers and editors, for
-     *                         example, checkboxes for Booleans.
-     * @param rowElementSetter A function which calls the appropriate setter for the
-     *                         column. The function is passed the element refererred
-     *                         to by the row (E) and the value from the
-     *                         TableCellEditor (Object). This should be casted
+     * @param columnClass      The class type the column deals with. This allows automatic setting
+     *                         of renderers and editors, for example, checkboxes for Booleans.
+     * @param rowElementSetter A function which calls the appropriate setter for the column. The
+     *                         function is passed the element refererred to by the row (E) and the
+     *                         value from the TableCellEditor (Object). This should be casted
      *                         appropriately.
-     * @param rowElementGetter A function which calls the appropriate getter for the
-     *                         column. The function is passed the element refererred
-     *                         to by the row (E) and is expected to return an
-     *                         Object.
-     * @param isColumnEditable If this is true, the column cells can be edited by
-     *                         the user using the column's TableCellEditor.
-     * @param columnCellEditor The TableCellEditor used to edit cells in this
-     *                         column. If this is null, swing will set the default
-     *                         TableCellEditor for the provided columnClass.
+     * @param rowElementGetter A function which calls the appropriate getter for the column. The
+     *                         function is passed the element refererred to by the row (E) and is
+     *                         expected to return an Object.
+     * @param isColumnEditable If this is true, the column cells can be edited by the user using the
+     *                         column's TableCellEditor.
+     * @param columnCellEditor The TableCellEditor used to edit cells in this column. If this is
+     *                         null, swing will set the default TableCellEditor for the provided
+     *                         columnClass.
      */
-    public ObjectTableColumn(String columnTitle, Class<?> columnClass, BiConsumer<E, Object> rowElementSetter,
-                             Function<E, Object> rowElementGetter, boolean isColumnEditable, TableCellEditor columnCellEditor) {
+    public ObjectTableColumn(String columnTitle, Class<?> columnClass,
+            BiConsumer<E, Object> rowElementSetter, Function<E, Object> rowElementGetter,
+            boolean isColumnEditable, TableCellEditor columnCellEditor) {
 
         this.columnTitle = columnTitle;
         this.columnClass = columnClass;
@@ -78,20 +75,20 @@ public class ObjectTableColumn<E> {
      * Returns the setter used by the TableColumnEditor for this column.
      *
      * @return The setter function in the format:
-     * <p>
-     * {@code (element, value) -> {}}
+     *         <p>
+     *         {@code (element, value) -> {}}
      */
     public BiConsumer<E, Object> getRowElementSetter() {
         return rowElementSetter;
     }
 
     /**
-     * Returns the getter used by the TableModel to tell it how to get the
-     * underlying element's attribute.
+     * Returns the getter used by the TableModel to tell it how to get the underlying element's
+     * attribute.
      *
      * @return The getter function, in the format:
-     * <p>
-     * {@code (element) -> {return (Object)value;}}
+     *         <p>
+     *         {@code (element) -> {return (Object)value;}}
      */
     public Function<E, Object> getRowElementGetter() {
         return rowElementGetter;
