+++ src/main/java/se/uog/table/ObjectTableModelAdaptor.java
@@ -1,10 +1,11 @@
 package se.uog.table;
 
-import javax.swing.*;
+import java.util.List;
+
+import javax.swing.DefaultListModel;
 import javax.swing.event.ListDataEvent;
 import javax.swing.event.ListDataListener;
 import javax.swing.table.AbstractTableModel;
-import java.util.List;
 
 /**
  * This class adapts an ObjectTableModel into a TableModel which can be used by
