+++ src/main/java/se/uog/training/TrainingTableModel.java
@@ -1,15 +1,16 @@
 package se.uog.training;
 
+import java.util.ArrayList;
+import java.util.List;
+
+import javax.swing.DefaultListModel;
+
 import se.uog.qualification.Qualification;
 import se.uog.table.ObjectTableColumn;
 import se.uog.table.ObjectTableColumnBuilder;
 import se.uog.table.ObjectTableListSelector;
 import se.uog.table.ObjectTableModel;
 import se.uog.teacher.Teacher;
-
-import javax.swing.*;
-import java.util.ArrayList;
-import java.util.List;
 
 public class TrainingTableModel implements ObjectTableModel<Training> {
